Potential fix for code scanning alert no. 52: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (size_t) to avoid overflow. This can be done by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done in size_t arithmetic, preventing overflow.

We will modify the line 220 to cast R + 1 to size_t before multiplying it by width.

+++ libwebp-1.0.3/src/utils/quant_levels_dec_utils.c
@@ -217,7 +217,7 @@ static int InitParams(uint8_t* const data, int width, int height, int stride,
                       int radius, SmoothParams* const p) {
   const int R = 2 * radius + 1;  // total size of the kernel
 
-  const size_t size_scratch_m = (R + 1) * width * sizeof(*p->start_);
+  const size_t size_scratch_m = (size_t)(R + 1) * width * sizeof(*p->start_);
   const size_t size_m =  width * sizeof(*p->average_);
   const size_t size_lut = (1 + 2 * LUT_SIZE) * sizeof(*p->correction_);
   const size_t total_size = size_scratch_m + size_m + size_lut;
